requirements.txt
google-generativeai
streamlit==1.28.0
python-dotenv

